Feat: Fetch specialties from API and create data fetching hooks
Introduces custom hooks (`useGetEspecialidades` and `useEspecialidadAgendas`) to manage fetching specialty and agenda data, including loading and error states.

Updates the home page to use `useGetEspecialidades` to dynamically populate the specialty selection dropdown in the appointment form, replacing the hardcoded list.

Fix bug that the hooks recive and date, but not the response